3 prilagodjavanje programa ulogovanom korisniku

+++ src/main/java/com/example/demo/controller/OperacijaController.java
@@ -13,15 +13,19 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.example.demo.dto.OperacijaDTO;
+import com.example.demo.model.AdministratorKlinike;
 import com.example.demo.model.Doktor;
 import com.example.demo.model.Klinika;
 import com.example.demo.model.Korisnik;
+import com.example.demo.model.Odsustvo;
 import com.example.demo.model.Operacija;
 import com.example.demo.model.Pregled;
 import com.example.demo.model.Sala;
 import com.example.demo.model.StatusOperacije;
 import com.example.demo.model.StatusPregleda;
 import com.example.demo.model.Termin;
+import com.example.demo.model.UlogaKorisnika;
+import com.example.demo.service.AdministratorKlinikeService;
 import com.example.demo.service.DoktorService;
 import com.example.demo.service.EmailService;
 import com.example.demo.service.KlinikaService;
@@ -56,17 +60,30 @@ public class OperacijaController {
 	
 	@Autowired
 	private KlinikaService klinikaService;
+	
+	@Autowired
+	private AdministratorKlinikeService adminKlinikeService;
 
 	
-	@RequestMapping(value = "/zahtevi", method=RequestMethod.GET) 
-	public ResponseEntity<List<Operacija>> zahtevi() {
+	@RequestMapping(value = "/zahtevi/{id}", method=RequestMethod.GET) 
+	public ResponseEntity<List<Operacija>> zahtevi(@PathVariable("id") Long identifikacija) {
+		Korisnik korisnik = korisnikService.findOne(identifikacija);
+		if (!korisnik.getUloga().equals(UlogaKorisnika.ADMIN_KLINIKE)) {
+			return new ResponseEntity<List<Operacija>>(HttpStatus.BAD_REQUEST);
+		}
+		AdministratorKlinike admin = adminKlinikeService.findByIdKorisnika(identifikacija);
 		List<Operacija> operacije = operacijaService.findByStatus(StatusOperacije.NERASPOREDJEN);
-		return new ResponseEntity<List<Operacija>>(operacije, HttpStatus.OK);
+		List<Operacija> operacijeKlinike = new ArrayList<Operacija>();
+		for (Operacija o : operacije) {
+			if(o.getPacijent().getKlinika().getId().equals(admin.getKlinika().getId())) {
+				operacijeKlinike.add(o);
+			}
+		}
+		return new ResponseEntity<List<Operacija>>(operacijeKlinike, HttpStatus.OK);
 	}
 	
 	@RequestMapping(value = "/rezervisi/{sala}/{operacija}/{doktori}", method = RequestMethod.GET)
 	public ResponseEntity<String> rezervisi(@PathVariable("sala") Long sala_id, @PathVariable("operacija") Long operacija_id, @PathVariable("doktori") String text) {
-		System.out.println("#######################");
 		System.out.println("sala" + sala_id + "operacija" + operacija_id + "doktori" + text);
 		Sala sala = salaService.findOne(sala_id);
 		Operacija operacija = operacijaService.findOne(operacija_id);
@@ -121,16 +138,32 @@ public class OperacijaController {
 		return new ResponseEntity<List<Operacija>>(operacije, HttpStatus.OK);
 	}
 	
-	@RequestMapping(value = "/obrisi/{id}", method = RequestMethod.DELETE)
-	public ResponseEntity<String> obrisi(@PathVariable("id") Long identifikacija) {
+	@RequestMapping(value = "/obrisi/{id}/{doktor_id}", method = RequestMethod.DELETE)
+	public ResponseEntity<String> obrisi(@PathVariable("id") Long identifikacija, @PathVariable("doktor_id") Long doktor_id) {
+		Korisnik korisnik = korisnikService.findOne(doktor_id);
+		if (!korisnik.getUloga().equals(UlogaKorisnika.LEKAR)) {
+			return new ResponseEntity<String>(HttpStatus.BAD_REQUEST);
+		}
 		Operacija operacija = operacijaService.findOne(identifikacija);
+		Doktor doktor = doktorService.findByIdKorisnik(doktor_id);
+		if (!operacija.getDoktori().contains(doktor)) {
+			return new ResponseEntity<String>(HttpStatus.BAD_REQUEST);
+		}
 		operacijaService.delete(operacija);
 		return new ResponseEntity<String>(HttpStatus.OK);
 	}
 	
-	@RequestMapping(value = "/izmeni", method = RequestMethod.PUT)
-	public ResponseEntity<OperacijaDTO> izmeni(@RequestBody OperacijaDTO operacijaDTO) {
+	@RequestMapping(value = "/izmeni/{id}", method = RequestMethod.PUT)
+	public ResponseEntity<OperacijaDTO> izmeni(@RequestBody OperacijaDTO operacijaDTO, @PathVariable("id") Long identifikacija) {
+		Korisnik korisnik = korisnikService.findOne(identifikacija);
+		if (!korisnik.getUloga().equals(UlogaKorisnika.LEKAR)) {
+			return new ResponseEntity<OperacijaDTO>(HttpStatus.BAD_REQUEST);
+		}
 		Operacija operacija = operacijaService.findOne(operacijaDTO.getId());
+		Doktor doktor = doktorService.findByIdKorisnik(identifikacija);
+		if (!operacija.getDoktori().contains(doktor)) {
+			return new ResponseEntity<OperacijaDTO>(HttpStatus.BAD_REQUEST);
+		}
 		operacija.setOpis(operacijaDTO.getOpis());
 		Operacija o = operacijaService.save(operacija);
 		return new ResponseEntity<OperacijaDTO>(new OperacijaDTO(o), HttpStatus.OK);
